'  Active display: Enable proximity sensor only when screen is off
This should ease people worried about cpu usage of the proximity
sensor while the screen is on.  The sensor listener will only be
registered when the screen goes off, and both active display and
pocket mode are enabled.  Once the screen turns on this listener
will be unregistered so no further updates are received when the
sensor values change.

 Active display: Use a default notificaiton icon.

In case there is an exception when trying to load the
notification icon we will use a default one.

Active display: Only fetch next notification if mNotification is null.